+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Missing images on collection page for venue selections
 
+### Changed
+
+- Let app know about its origin instead of its graphql endpoint
+
 ## [1.0.0] - 2022-03-08
 
 ### Fixed